Use separate file for connected layout components
