+++ src/Layout/ReduxNavBar.js
@@ -1,0 +1,9 @@
+import { connect } from 'react-redux';
+import NavBar from '../components/layout/NavBar';
+
+const mapStateToProps = (state) => ({
+  userAccess: state.userAccess.access,
+});
+
+const ReduxNavBar = connect(mapStateToProps)(NavBar);
+export default ReduxNavBar;
