+++ src/Layout/ReduxTopBar.js
@@ -1,0 +1,16 @@
+import { connect } from 'react-redux';
+import TopBar from '../components/layout/TopBar';
+import { logoutAPI } from '../actions';
+
+const resourcePath = '/services/sheepdog/submission/project';
+const mapStateToProps = (state) => ({
+  username: state.user.username,
+  isAdminUser: state.user.authz?.[resourcePath]?.[0].method === '*',
+});
+
+const mapDispatchToProps = (dispatch) => ({
+  onLogoutClick: () => dispatch(logoutAPI()),
+});
+
+const ReduxTopBar = connect(mapStateToProps, mapDispatchToProps)(TopBar);
+export default ReduxTopBar;
